requirements.txt
# requirements.txt
Django~=5.0
celery[redis]
pandas
numpy
python-dateutil
prophet
WeasyPrint
mysqlclient
python-decouple